Add leader discovery for VeniceParentHelixAdmin, remove the support for child controller
